Add files via upload
finalized creation of grade selector buttons on project page, this component will likely be reused on other pages. Initialized and connected script.js into the portfolio page and added script to allow the div for each grade to auto hide until the button is clicked for that grade, then it will display the div with that grades portfolios.